Prototype implementation of remote execution
This code extends adds a RemoteWorkExecutor interface for remote execution.
This is implemented by RestWorkExecutor that submits the remote work as
a blocking REST call.

Also adding a RestBuildWorker which is a Jetty server with a Servlet to handle
build request. It then runs the build request synchronously. It redirects the
work request to MemcacheWorkExecutor which runs the work locally.

Change-Id: Ie714cd29d95a77193798e8d9b58ddd92863cf1bf