Remote execution now works for a simple c++ compilation
Fixing up the whole logic for saving input files to cache, writing
cached output files to file system and save the to cache.

RestBuildWorker will write the files from cache to disk and then run
the work synchronously. It then removes the directory after the work.

Once the work is done it sends a RemoteWorkResponse message (in JSON)
to the caller.

Change-Id: I246f350d34f163d23f41f95e1f52bc6e3cea8691