Implement distributed caching for Bazel
This patch implements distributed caching for Bazel using the JCache API.
The JCache API is used as a key value store that stores content of files
indexed by the digest of the file. The cache also stores the list of files
for an action. The key in this case is the digest from the key of the action
and the list of files.

The caching backend is Hazelcast which is a memory only distribued cache.
Unfortunately adding this dependency to Bazel makes the final binary 7MB
bigger.

In this change I also added the interface for remote execution. The
implementation will be added in a subsequent patch.

This change is only the first in a series of changes related to distributed
caching and remote execution. I plan to revise the APIs and implementation
in subsequent changes.