Prototype for remote execution using gRPC and Netty transport
This change implements a remote worker that executes work (build or test).
Bazel will be a client of the remote worker. The communication uses gRPC
and Netty as transport.

A single remote worker has little advantage over running locally. Additional
infrastructure is needed to run workers on multiple machines and distributing
the work among them.

This change provides the basic building blocks for a distributed build farm.

Change-Id: If7d285444ef42a6823b59443af17b61b04b9ce6a